Refactory de métodos antes ao onCreate
