+++ app/src/main/java/br/com/biexpert/bicm/dto/TaskModel.kt
@@ -1,4 +1,4 @@
-package br.com.biexpert.bicm
+package br.com.biexpert.bicm.dto
